(PC-16323) refactor(pageHeader): remove old PageHeader
