+++ src/features/bookings/pages/EndedBookings.tsx
@@ -35,8 +35,12 @@ export const EndedBookings: React.FC = () => {
 
   return (
     <React.Fragment>
-      <PageHeader title={t`Réservations terminées`} />
-      <Spacer.TopScreen />
+      <PageHeader
+        title={t`Réservations terminées`}
+        background="primary"
+        size="small"
+        withGoBackButton
+      />
       <FlatList
         contentContainerStyle={contentContainerStyle}
         data={bookings?.ended_bookings || []}
