+++ src/features/profile/pages/ChangeEmail/ChangeEmail.tsx
@@ -101,13 +101,17 @@ export function ChangeEmail() {
 
   return (
     <React.Fragment>
-      <PageHeader title={t`Modifier mon e-mail`} />
-      <Spacer.TopScreen />
+      <PageHeader
+        title={t`Modifier mon e-mail`}
+        background="primary"
+        size="small"
+        withGoBackButton
+      />
       <StyledScrollView
         ref={scrollRef}
         contentContainerStyle={getScrollViewContentContainerStyle(keyboardHeight)}
         keyboardShouldPersistTaps="handled">
-        <Spacer.Column numberOfSpaces={18} />
+        <Spacer.Column numberOfSpaces={6} />
         {hasCurrentEmailChange ? (
           <React.Fragment>
             <AlreadyChangedEmailDisclaimer />
