+++ src/features/search/pages/SearchFilter.tsx
@@ -45,11 +45,16 @@ export const SearchFilter: React.FC = () => {
 
   return (
     <Container>
-      <PageHeader title={t`Filtrer`} RightComponent={ReinitializeFilters} />
+      <PageHeader
+        title={t`Filtrer`}
+        RightComponent={ReinitializeFilters}
+        background="primary"
+        size="small"
+        withGoBackButton
+      />
       <React.Fragment>
         <StyledScrollView ref={scrollViewRef} scrollEnabled={scrollEnabled}>
-          <Spacer.TopScreen />
-          <Spacer.Column numberOfSpaces={16} />
+          <Spacer.Column numberOfSpaces={4} />
 
           {/* Localisation */}
           <Section.Location />
